requirements.txt
pandas
openpyxl
smtplib
email
python-dotenv
logging
